refactor: streamline schedule components and enhance date formatting utilities
