+++ package.json
@@ -30,7 +30,6 @@
     "react": "19.2.4",
     "react-dom": "19.2.4",
     "react-hook-form": "^7.73.1",
-    "recharts": "^3.8.0",
     "server-only": "^0.0.1",
     "shadcn": "^4.5.0",
     "tailwind-merge": "^3.5.0",
