--- src/features/schedule/lib/formatDateHeader.ts
@@ -1,7 +1,0 @@
-export function formatDateHeader(date: Date, locale: string): string {
-  return new Intl.DateTimeFormat(locale, {
-    weekday: "long",
-    month: "long",
-    day: "numeric",
-  }).format(date);
-}
